applied patch
plus an edit to fix username not updating in the modal after oauth finishes
