+++ client/src/utilities.js
@@ -116,3 +116,27 @@ export function exportTextFile({ content, contentType = "text/csv", fileName, ch
   dl.download = fileName;
   dl.click();
 }
+
+export async function showAuthPopup(authEndpoint, setUser) {
+  const width = 600;
+  const height = 600;
+  const left = window.innerWidth / 2 - width / 2;
+  const top = window.innerHeight / 2 - height / 2;
+
+  const popup = window.open(
+    authEndpoint,
+    "",
+    `toolbar=no, location=no, directories=no, status=no, menubar=no,
+      scrollbars=no, resizable=no, copyhistory=no, width=${width},
+      height=${height}, top=${top}, left=${left}`
+  );
+
+  const loop = setInterval(async () => {
+    if (popup.closed) {
+      clearInterval(loop);
+      const userData = await get("/api/whoami");
+      setUser(userData);
+    }
+  }, 50);
+  return loop;
+}
